feat: Add validation and return methods in useHandler
The code changes in `useHandler.jsx` include adding validation for publication data and returning the methods to be used in the component. This commit message suggests that the purpose of the changes is to improve the functionality of the `useHandler` hook by adding validation for publication data and returning the necessary methods.