feat: Add estado field to DetallesPublicacion and PostPublicationForm
The code changes add a new field called "estado" to the components DetallesPublicacion and PostPublicationForm. This field allows users to specify the state of the publication. The estado field is optional and will be used to indicate whether the publication is available, sold, or reserved. This enhancement improves the functionality and flexibility of the application.